chg: move Token type to storage/local/index.ts
