+++ app/storage/index.ts
@@ -4,10 +4,10 @@ import { create, type StoreApi } from "zustand"
 import { useShallow } from "zustand/react/shallow"
 
 import { StorageAction } from "~background/messages/storage"
-import type { Token } from "~packages/account"
 import { PasskeyAccount } from "~packages/account/PasskeyAccount"
 import type { UserOperationData } from "~packages/bundler"
 import number from "~packages/util/number"
+import type { Token } from "~storage/local"
 import {
   UserOperationStatus,
   type State,
